Split structure module into three
